+++ modules/network/sonic.ts
@@ -140,6 +140,29 @@ export const sonicNetworkConfig: NetworkConfig = {
         {
             name: 'sync-erc4626-unwrap-rate',
             interval: (env.DEPLOYMENT_ENV as DeploymentEnv) === 'canary' ? every(60, 'minutes') : every(20, 'minutes'),
+        },        {
+            name: 'add-pools-v3',
+            interval: every(2, 'minutes'),
+        },
+        {
+            name: 'sync-pools-v3',
+            interval: every(30, 'seconds'),
+        },
+        {
+            name: 'sync-join-exits-v3',
+            interval: every(1, 'minutes'),
+        },
+        {
+            name: 'sync-swaps-v3',
+            interval: every(1, 'minutes'),
+        },
+        {
+            name: 'sync-snapshots-v3',
+            interval: every(10, 'minutes'),
+        },
+        {
+            name: 'sync-hook-data',
+            interval: every(1, 'hours'),
         },
     ],
 };