+++ test/factories/prismaPool.factory.ts
@@ -6,7 +6,6 @@ import { Chain, PrismaPoolType } from '@prisma/client';
 import { prismaPoolDynamicDataFactory } from './prismaPoolDynamicData.factory';
 import { LiquidityManagement } from '../../modules/sor/types';
 
-
 class PrismaPoolFactory extends Factory<PrismaPoolAndHookWithDynamic> {
     stable(amp?: string) {
         return this.params({ type: PrismaPoolType.STABLE, typeData: { amp: amp ?? '10' } });
@@ -22,6 +21,7 @@ export const prismaPoolFactory = PrismaPoolFactory.define(({ params }) => {
         enableDonation: false,
         enableRemoveLiquidityCustom: false,
     };
+    const chain = params?.chain || Chain.SEPOLIA;
 
     return {
         id: poolAddress,
@@ -34,14 +34,14 @@ export const prismaPoolFactory = PrismaPoolFactory.define(({ params }) => {
         pauseManager: createRandomAddress(),
         poolCreator: createRandomAddress(),
         factory: createRandomAddress(),
-        chain: Chain.SEPOLIA,
+        chain,
         version: 1,
         protocolVersion: 3,
         typeData: {},
         categories: [],
         createTime: 1708433018,
-        dynamicData: prismaPoolDynamicDataFactory.build({ id: poolAddress, chain: params?.chain || Chain.SEPOLIA }),
-        tokens: prismaPoolTokenFactory.buildList(2),
+        dynamicData: prismaPoolDynamicDataFactory.build({ id: poolAddress, chain }),
+        tokens: prismaPoolTokenFactory.buildList(2, { chain }),
         hookId: null,
         hook: hook,
         liquidityManagement: liquidityManagement,